refactor(shared): move app types to shared folder
