+++ client/src/hooks/useLoginState.tsx
@@ -1,6 +1,6 @@
 import {useOutletContext} from 'react-router-dom';
 
-import type {LoginState} from '../types/app';
+import type {LoginState} from '../shared/types/appTypes';
 
 export function useLoginState() {
   return useOutletContext<LoginState>();
