+++ client/src/routes/app/App.tsx
@@ -1,15 +1,14 @@
 import React from 'react';
-import {Outlet, Link} from 'react-router-dom';
-
+import {Outlet} from 'react-router-dom';
 import {useLocalStorage} from '../../hooks';
 
-import type {LoginState} from '../../types/app';
+import type {LoginState} from '../../shared/types';
 
 export function App() {
   const [storedValue, setValue] = useLocalStorage('loginStatus', false);
 
   return (
-    <div className="App">
+    <div className="app">
       <Outlet context={[storedValue, setValue] satisfies LoginState} />
     </div>
   );
