+++ client/src/shared/types/index.tsx
@@ -1,1 +1,2 @@
+export type {LoginState} from './appTypes';
 export type {NavigationType} from './navigationTypes';
